[Group 13] Customer products correction

+++ back-end/src/database/migrations/20220711220859-create-products.js
@@ -14,7 +14,7 @@ module.exports = {
         allowNull: false,
       },
       price: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL(4,2),
         allowNull: false,
       },
       url_image: {
